Flytt logikk for å slå sammen overlappende eller påfølkgende ytelses perioder for justere tilgangen på metodene
