+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/grunnlag/VilkårperioderGrunnlag.kt
@@ -2,9 +2,6 @@ package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag
 
 import no.nav.tilleggsstonader.kontrakter.aktivitet.Kilde
 import no.nav.tilleggsstonader.kontrakter.aktivitet.StatusAktivitet
-import no.nav.tilleggsstonader.kontrakter.felles.Periode
-import no.nav.tilleggsstonader.kontrakter.felles.mergeSammenhengende
-import no.nav.tilleggsstonader.kontrakter.felles.overlapperEllerPåfølgesAv
 import no.nav.tilleggsstonader.kontrakter.ytelse.TypeYtelsePeriode
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.infrastruktur.database.Sporbar
@@ -14,8 +11,6 @@ import org.springframework.data.relational.core.mapping.Table
 import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
-import kotlin.collections.plus
-import kotlin.collections.sortedBy
 import no.nav.tilleggsstonader.kontrakter.ytelse.EnsligForsørgerStønadstype as EnsligForsørgerStønadstypeKontrakter
 
 data class VilkårperioderGrunnlag(
@@ -65,45 +60,7 @@ data class PeriodeGrunnlagYtelse(
     val fom: LocalDate,
     val tom: LocalDate?,
     val ensligForsørgerStønadstype: EnsligForsørgerStønadstype? = null,
-) {
-    companion object {
-
-        fun List<PeriodeGrunnlagYtelse>.slåSammenOverlappendeEllerPåfølgende(): List<PeriodeGrunnlagYtelse> {
-            val (perioderMedTom, perioderUtenTom) = this.partition { it.tom != null }
-
-            val sammenslåttePerioder = perioderMedTom
-                .map { PeriodeGrunnlagYtelseHolder(it) }
-                .sortedWith(compareBy({ it.ytelse.type }, { it }))
-                .mergeSammenhengende(
-                    skalMerges = { v1, v2 -> v1.kanSlåsSammen(v2) },
-                    merge = { v1, v2 -> v1.slåSammen(v2) },
-                )
-                .map { it.ytelse }
-
-            return (sammenslåttePerioder + perioderUtenTom).sortedBy { it.fom }
-        }
-    }
-}
-
-data class PeriodeGrunnlagYtelseHolder(
-    val ytelse: PeriodeGrunnlagYtelse,
-) : Periode<LocalDate> {
-    override val fom: LocalDate
-        get() = ytelse.fom
-    override val tom: LocalDate
-        get() = ytelse.tom ?: error("Mangler tom")
-
-    fun slåSammen(other: PeriodeGrunnlagYtelseHolder): PeriodeGrunnlagYtelseHolder =
-        PeriodeGrunnlagYtelseHolder(
-            ytelse.copy(
-                fom = minOf(ytelse.fom, other.ytelse.fom),
-                tom = maxOf(ytelse.tom!!, other.ytelse.tom!!),
-            ),
-        )
-
-    fun kanSlåsSammen(other: PeriodeGrunnlagYtelseHolder): Boolean =
-        ytelse.type == other.ytelse.type && ytelse.ensligForsørgerStønadstype == other.ytelse.ensligForsørgerStønadstype && overlapperEllerPåfølgesAv(other)
-}
+)
 
 data class HentetInformasjon(
     val fom: LocalDate,
