+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/SlåSammenPeriodeGrunnlagYtelseStepDefinition.kt
@@ -14,7 +14,7 @@ import no.nav.tilleggsstonader.sak.cucumber.parseValgfriDato
 import no.nav.tilleggsstonader.sak.cucumber.parseValgfriEnum
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.EnsligForsørgerStønadstype
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.PeriodeGrunnlagYtelse
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.PeriodeGrunnlagYtelse.Companion.slåSammenOverlappendeEllerPåfølgende
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.grunnlag.SlåSammenPeriodeGrunnlagYtelseUtil.slåSammenOverlappendeEllerPåfølgende
 import org.assertj.core.api.Assertions.assertThat
 
 enum class PeriodeGrunnlagYtelseNøkler(
